+++ CHANGELOG.md
@@ -7,13 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2019-07-01
+
+### Added
+
+- Add `initWithWithCVPixelBuffer:glTextureCache:` and `initWithSize:glTextureCache:` in *GLIRenderTarget*.
+- Add `glTextureCache` in *GLIRenderTarget*.
+- Add `initWithWithCVPixelBuffer:glTextureCache:mtlTextureCache:` and `initWithSize:glTextureCache:mtlTextureCache:` in *GLIMetalRenderTarget*.
+- Add `mtlTextureCache` in *GLIMetalRenderTarget*.
+- Add `setTextureParameters` in *GLITexture*.
+
+### Deprecated
+
+- Deprecate `upload` in *GLITexture* for incorrect API naming reason.
+
 ## [1.2.0] - 2019-06-27
 
-## Changed
+### Changed
 
 - Change the type of `output` from `GLIRenderTarget *` to `__kindof GLIRenderTarget *` in *GLIRenderer* to compatible with *GLIMetalRenderTarget*.
 
-## Added
+### Added
 
 - Add *GLIMetalRenderTarget*.
 - Add *GLIMetalTextureCache*.